src: Make arch-syscall-check work in VPATH build
The aim of arch-syscall-check test is to check for syscalls
missing implementation. It does so by comparing two files:

  1) src/syscalls.csv
  2) include/seccomp-syscalls.h

However, due to use of relative paths these files are not found
when doing a VPATH build. But, we can re-use an idea from GNU
coreutils and get an absolute path to the source dir. All that's
needed then is to prefix those two paths with the source dir
path.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Acked-by: Tom Hromatka <tom.hromatka@oracle.com>
Signed-off-by: Paul Moore <paul@paul-moore.com>
(cherry-picked from commit 744c9a897b74ad66d065791593e25a05e4b6f6a1)